feat: Melhorando os aspectos das telas de fronteiras eas adaptando para uma versao proxima do final e concluindo as DAOs

+++ src/main/java/model/Agencia.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Agencia {
 	
-	private int codigo;
+	private long codigo;
 	private String nome;
 	private String cep;
 	private String cidade;
